add node builing mobule
you can now build a module via node. it is simple and have a web interface. whit a drag and drop panel.

you can easy made new node for the builder, look in the ´nodeStandardModules´, folder. where each file can have a list of nodes. then when loading the system, it can be used in the node builder.